+++ mlxtend/feature_selection/__init__.py
@@ -5,5 +5,7 @@
 from .feature_sel import ColumnSelector
 from .sequential_backward_select import SBS
 from .sequential_forward_select import SFS
+from .sequential_floating_backward_select import SFBS
+from .sequential_floating_forward_select import SFFS
 
-__all__ = ["ColumnSelector", "SBS", "SFS"]
+__all__ = ["ColumnSelector", "SBS", "SFS", "SFBS", "SFFS"]