+++ mlxtend/feature_selection/sequential_floating_backward_select.py
@@ -1,6 +1,6 @@
 # Sebastian Raschka, 2015
 # Function for sequential feature selection via
-# Sequential Backward Selection (SBS)
+# Sequential Floating Backward Selection (SFBS)
 # mlxtend Machine Learning Library Extensions
 
 from sklearn.base import clone
@@ -9,11 +9,11 @@ from sklearn.base import MetaEstimatorMixin
 from itertools import combinations
 from sklearn.cross_validation import cross_val_score
 import numpy as np
+from collections import deque
 import sys
 
-
-class SBS(BaseEstimator, MetaEstimatorMixin):
-    """ Sequential Backward Selection for feature selection.
+class SFBS(BaseEstimator, MetaEstimatorMixin):
+    """ Sequential Floating Backward Selection for feature selection.
 
     Parameters
     ----------
@@ -31,6 +31,12 @@ class SBS(BaseEstimator, MetaEstimatorMixin):
 
     cv : int (default: 5)
       Number of folds in StratifiedKFold.
+
+    max_iter: int (default: -1)
+      Terminate early if number of `max_iter` is reached.
+
+    skip_if_stuck: bool (default: True)
+      If `True`, skips conditional exclusion step if stuck.
 
     n_jobs : int (default: 1)
       The number of CPUs to use for cross validation. -1 means 'all CPUs'.
@@ -57,53 +63,85 @@ class SBS(BaseEstimator, MetaEstimatorMixin):
     >>> X = iris.data
     >>> y = iris.target
     >>> knn = KNeighborsClassifier(n_neighbors=4)
-    >>> sbs = SBS(knn, k_features=2, scoring='accuracy', cv=5)
-    >>> sbs = sbs.fit(X, y)
-    >>> sbs.indices_
-    (0, 3)
-    >>> sbs.transform(X[:5])
-    array([[ 5.1,  0.2],
-           [ 4.9,  0.2],
-           [ 4.7,  0.2],
-           [ 4.6,  0.2],
-           [ 5. ,  0.2]])
+    >>> sfbs = SFBS(knn, k_features=2, scoring='accuracy', cv=5)
+    >>> sfbs = sfbs.fit(X, y)
+    >>> sfbs.indices_
+    (2, 3)
+    >>> sfbs.transform(X[:5])
+    array([[ 1.4,  0.2],
+           [ 1.4,  0.2],
+           [ 1.3,  0.2],
+           [ 1.5,  0.2],
+           [ 1.4,  0.2]])
 
-    >>> print('best score: %.2f' % sbs.k_score_)
-    best score: 0.96
+    >>> print('best score: %.2f' % sfbs.k_score_)
+    best score: 0.97
 
     """
     def __init__(self, estimator, k_features, print_progress=True,
-                 scoring='accuracy', cv=5, n_jobs=1):
+                 scoring='accuracy', max_iter=-1, cv=5,
+                 skip_if_stuck=True, n_jobs=1):
         self.scoring = scoring
-        self.estimator = estimator#clone(estimator)
+        self.estimator = estimator #clone(estimator)
         self.cv = cv
+        self.max_iter = max_iter
         self.k_features = k_features
+        self.skip_if_stuck = skip_if_stuck
         self.print_progress = print_progress
         self.n_jobs = n_jobs
 
     def fit(self, X, y):
 
         dim = X.shape[1]
-        self.indices_ = tuple(range(dim))
+        orig = tuple(range(dim))
+        self.indices_ = orig
         self.subsets_ = [self.indices_]
         cv_score = self._calc_score(X, y, self.indices_)
         self.scores_ = [cv_score.mean()]
 
-        while dim > self.k_features:
-            scores = []
-            subsets = []
+        if self.skip_if_stuck:
+            sdq = deque(maxlen=4)
+        else:
+            sdq = deque(maxlen=0)
 
+        cnt = 0
+        while dim > self.k_features and cnt != self.max_iter:
+            scores_1, scores_2 = [], []
+            subsets_1, subsets_2 = [], []
+
+            # step 1: exclusion
             for p in combinations(self.indices_, r=dim-1):
                 cv_score = self._calc_score(X, y, p)
-                scores.append(cv_score.mean())
-                subsets.append(p)
+                scores_1.append(cv_score.mean())
+                subsets_1.append(p)
 
-            best = np.argmax(scores)
-            self.indices_ = subsets[best]
-            self.subsets_.append(self.indices_)
-            dim -= 1
+            best_1 = np.argmax(scores_1)
+            self.indices_ = subsets_1[best_1]
+            excluded = set(orig) - set(self.indices_)
 
-            self.scores_.append(scores[best])
+            # step 2: conditional inclusion
+            best_2_score = -1.0
+            if not len(sdq) == 4 or (sdq[0] != sdq[2] or sdq[1] != sdq[3]):
+                set_indices = set(self.indices_)
+                for i in excluded:
+                    test_subset = tuple(sorted(set_indices | {i}))
+                    cv_score = self._calc_score(X, y, test_subset)
+                    scores_2.append(cv_score.mean())
+                    subsets_2.append(test_subset)
+
+                best_2 = np.argmax(scores_2)
+                best_2_score = scores_2[best_2]
+
+
+            if best_2_score > scores_1[best_1]:
+                self.indices_ = subsets_2[best_2]
+            else:
+                self.subsets_.append(self.indices_)
+                self.scores_.append(scores_1[best_1])
+                dim -= 1
+
+            sdq.append(self.indices_)
+            cnt += 1
 
             if self.print_progress:
                 sys.stderr.write('\rFeatures: %d/%d' % (len(self.indices_), self.k_features))