+++ tests/tests_feature_selection/test_sequential_forward_select.py
@@ -12,8 +12,19 @@ def test_Iris():
 
     knn = KNeighborsClassifier(n_neighbors=4)
 
-    sfs = SFS(knn, k_features=2, scoring='accuracy', cv=5)
+    sfs = SFS(knn, k_features=2, scoring='accuracy', cv=5, print_progress=False)
     sfs.fit(X, y)
 
     assert(sfs.indices_ == (2, 3))
     assert(round(sfs.k_score_, 2) == 0.97 )
+
+
+def test_selects_all():
+    from sklearn.neighbors import KNeighborsClassifier
+    from mlxtend.data import wine_data
+
+    X, y = wine_data()
+    knn = KNeighborsClassifier(n_neighbors=4)
+    sfs = SFS(knn, k_features=13, scoring='accuracy', cv=3, print_progress=False)
+    sfs.fit(X, y)
+    assert(len(sfs.indices_) == 13)