+++ docs/sources/CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.3.0
 
-- Sequential Feature Selection algorithms: [SFS](./docs/feature_selection/sequential_forward_selection/), [SFFS](./docs/feature_selection/sequential_floating_forward_selection/), and [SFBS](./docs/feature_selection/sequential_floating_backward_selection/)
+
+- Sequential Feature Selection algorithms: [SFS](http://rasbt.github.io/mlxtend/docs/feature_selection/sequential_forward_selection/), [SFFS](http://rasbt.github.io/mlxtend/docs/feature_selection/sequential_floating_forward_selection/), and [SFBS](http://rasbt.github.io/mlxtend/docs/feature_selection/sequential_floating_backward_selection/)
 
 ### 0.2.9
 - Changed `regularization` & `lambda` parameters in `LogisticRegression` to single parameter `l2_lambda`